cloudserver/cloudserver/urls.py
"""
URL configuration for cloudserver project.

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/4.2/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  path('', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
Including another URLconf
    1. Import the include() function: from django.urls import include, path
    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
"""

from django.contrib import admin
from django.urls import include, path
# Ensure this incorrect import is removed or commented out:
# from core.views import generate_jwt_for_guacamole
from core.views import GuacamoleAdminAPIView, GenerateGuacamoleTokenViewExpert
from core.views import LoginView
from rest_framework import permissions
from drf_yasg import openapi
from drf_yasg.views import get_schema_view
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView
from core.views import GameList

schema_view = get_schema_view(
   openapi.Info(
      title="API Documentation",
      default_version='v1',
      description="API documentation for MyApp",
   ),
   public=True,
   permission_classes=(permissions.AllowAny,),
)

urlpatterns = [
    path("admin/", admin.site.urls),
    path("", include("core.urls")),  # Add this line
    path('swagger/', schema_view.with_ui('swagger', cache_timeout=0), name='schema-swagger-ui'),
    path('api/admin/guacamole-connections/', GuacamoleAdminAPIView.as_view(), name='admin-guacamole-connections'),  # New path for the admin view
    path('api/generate-guacamole-token-expert/', GenerateGuacamoleTokenViewExpert.as_view(), name='generate_guacamole_token'),
    path('api/login/', LoginView.as_view(), name='login'),
    path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('api/games/', GameList.as_view(), name='game-list'),
]

